front end
Changed the JSP files to make them look a little fancier and added in a page so the index page is separate from the add new address files. We will need to make sure that we get to the changes for the various links and buttons values in the servlets from what I changed. (make sure that all the mapping is correct)